% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_betas.R
\name{bumpsBeta}
\alias{bumpsBeta}
\title{BumpsBeta}
\usage{
bumpsBeta(seed, n = 64)
}
\arguments{
\item{seed}{intger, set a seed}

\item{n}{side length of the quadratic image}
}
\value{
bumpy image as matrix
}
\description{
Returns a bumpy image
}
\examples{
require(magrittr)
### divide by 50 and decrease resolution to obtain a bumpy 32x32 image.
bumpsBeta(123) \%>\%
divide_by(50) \%>\%
 decrease_reolution() \%>\%
 plot_coefficient_image()

}
