% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_struc_mat.R
\name{calc_struc_mat}
\alias{calc_struc_mat}
\title{Calculates structure matrix for sx objects}
\usage{
calc_struc_mat(L, neighbours = c("2dfirst", "2dsecond", "2dallfirst",
  "3dfirst", "3dsecond", "3dallfirst"))
}
\arguments{
\item{L}{integer vector, defining dimension of image}

\item{neighbours}{indicates the used neighbourhood. Currently for 2d/3d images}
}
\value{
structure matrix for SOIR ()
}
\description{
Calculates structure matrix with a given dimension and neighbourhood assumption
}
