% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_SOIR.R
\name{decrease_reolution}
\alias{decrease_reolution}
\title{Decrease resolution of an image}
\usage{
decrease_reolution(image)
}
\arguments{
\item{image}{image represented as matrix wirh dimension (d x d), Note that d must be even}
}
\value{
matrix of dimension (d/2 x d/2)
}
\description{
Decreasing resolution of an image from dimension (d x d) to (d/2 x d/2)
}
\examples{
See example from bumpsBeta()
}
