% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_SOIR.R
\name{get_beta}
\alias{get_beta}
\title{Extract beta coeffiscients from sarim output}
\usage{
get_beta(model, intercept = FALSE, burnin = 50, reduce = TRUE,
  reduce_with = mean)
}
\arguments{
\item{model}{list, output list from sarim function}

\item{intercept}{logical indicating if sarim was calculated with intercept}

\item{burnin}{integer, controlls how many iterations at start should be deleted}

\item{reduce}{logical, controlls wether the output should be reduced to 1 number per beta}

\item{reduce_with}{function, defines the function to reduce output, e.g. mean}
}
\value{
either a vector, the mean of each beta or a matrix
}
\description{
Extract beta coeffiscients from sarim output
}
\examples{
beta <- rnorm(32*32)
y <- ims  \%*\% beta + rnorm(100, sd = 5)
Sarim::sarim(y ~ SOIR(ims, neighbours = "2dallfirst", add_diag = 0.1), nIter = 10) \%>\%
get_beta(burnin = 2) \%>\%
plot_coefficient_image
}
