% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOIR.R
\name{plot.sx}
\alias{plot.sx}
\title{Plot function of sx object}
\usage{
\method{plot}{sx}(x, col = grey.colors(100), ...)
}
\arguments{
\item{x}{A sx object}

\item{full}{colors for plot}
}
\value{
Null
}
\description{
plots the neighbourhood structue of sx objects
}
