% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefficient_image.R
\name{plot_coefficient_image}
\alias{plot_coefficient_image}
\title{Plotting a coefficient image}
\usage{
plot_coefficient_image(im, title = NULL,
  col = colorRampPalette(c("black", "white"))(100), barh = 12,
  breaks = 5, round = 3)
}
\arguments{
\item{im}{image given as matrix}

\item{title}{String, title of the image, default is NULL (no title)}

\item{col}{color palette of the image, default is color interplotaion from black to white}

\item{barh}{numeric, only nedded to adjust the height of the legend bar}

\item{breaks}{number, of breaks for the legend}

\item{round}{integer, indicates, how many decimal digits are use in the legend}
}
\value{
A ggplot object
}
\description{
Plotting a coefficient image using ggplot
}
