% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOIR.R
\name{print.sx}
\alias{print.sx}
\title{Print function for sx objects}
\usage{
\method{print}{sx}(x, full = F)
}
\arguments{
\item{x}{A sx object}

\item{full}{Logical, indicating if output should be full}
}
\value{
Null
}
\description{
Since the output of Sarim::sx() is a very long list, a short version is
implemented. The option full = TRUE can be used, to get the full printet list object.
}
\details{
Easy print function to get nice readable output for sx-object
}
