% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_SOIR.R
\name{set_dim}
\alias{set_dim}
\title{Setting dimensions within a pipe}
\usage{
set_dim(x, dim)
}
\arguments{
\item{x}{Numeric matrix}

\item{dim}{Vector containing the dimension to set}
}
\value{
same matrix with new dimension
}
\description{
Setting dimensions in pipe-like environment from e.g. magrittr
}
\examples{
require(magrittr)
rnorm(100) \%>\% set_dim(c(20,5)) \%>\% head
}
