% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOIR.R
\name{SOIR}
\alias{SOIR}
\title{Scalar-on-Image regression in STAR framework using sarim}
\usage{
SOIR(images, dimension = rep(sqrt(ncol(images)), 2),
  neighbours = c("2dfirst", "2dsecond", "2dallfirst", "3dfirst",
  "3dsecond", "3dallfirst"), solver = c("rue", "lanczos"),
  demean = FALSE, add_diag = NULL, ...)
}
\arguments{
\item{images}{The vectorized images (columnwise)}

\item{dimension}{Vector which defines the dimension of the images}

\item{neighbours}{String defining the type of neighbourhood of underlying GMRF}

\item{solver}{String defining the solve which should be used}

\item{demean}{Logical, definign wether the mena of images should be substraced}

\item{add_diag}{numeric value, default is NULL. For numeric stability one can
add a value to main diagonal to structure matrix}

\item{...}{Further arguments for the underlying function \code{Sarim::sx()},
most relevant are ka_a, ka_b and ka_start}
}
\value{
sx-object with images, structure matrix and so on.
}
\description{
Returns a sx-object for Scalar-on-Image regression within sarim function
}
\examples{

### simple call:
ims <- matrix(1:(100*32*32), nrow = 100, ncol = 32*32)
SOIR(ims)

### visualization of the structure matrix
plot(SOIR(ims))

### Call within Sarim:
beta <- rnorm(32*32)
y <- ims  \%*\% beta + rnorm(100, sd = 5)
mod <- Sarim::sarim(y ~ SOIR(ims, add_diag = 0.1, ), nIter = 10)

}
