% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_betas.R
\name{sparseBeta}
\alias{sparseBeta}
\title{SparseBeta}
\usage{
sparseBeta(x1, x2, random = FALSE)
}
\arguments{
\item{x1}{sequence between 0 and 1}

\item{x2}{see x1}

\item{random}{logical: Random generated?}
}
\value{
sparse image as matrix
}
\description{
Returns a sparse image
}
\examples{
require(magrittr)
grid <- seq(0,1, len = 64)
sparseBeta(grid,grid) \%>\% plot_coefficient_image()
}
